# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  errno.Errno:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  user.InfoResponse:
    properties:
      avatar:
        type: string
      degree:
        type: string
      doing:
        type: integer
      done:
        type: integer
      earning:
        type: string
      email:
        type: string
      gender:
        type: string
      nick_name:
        type: string
      phone_num:
        type: string
    type: object
  user.UpdateInfoRequest:
    properties:
      degree:
        type: string
      nick_name:
        type: string
      old_password:
        type: string
      password:
        type: string
      phone_num:
        type: string
    type: object
host: 119.3.2.168:2333
info:
  contact:
    email: 1903180340@qq.com
    name: Wishiforpeace
  description: 纸贵API
  termsOfService: http://swagger.io/terrms/
  title: Zhigui API
  version: 1.0.0
paths:
...
  /user/avatar:
    post:
      consumes:
      - application/json
      description: 修改用户头像
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"上传成功","data":map[string]interface{"url","path","sha"}}'
          schema:
            type: string
        "400":
          description: 上传失败,请检查token与其他配置参数是否正确
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: 修改头像
      tags:
      - user
  /user/info:
    get:
      consumes:
      - application/json
      description: 得到用户所有的个人信息
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":{"username":"kong"}}'
          schema:
            $ref: '#/definitions/user.InfoResponse'
      summary: 得到用户信息
      tags:
      - user
    post:
      consumes:
      - application/json
      description: 修改用户所有的个人信息
      parameters:
      - description: 需要修改对内容
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/user.UpdateInfoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"NickName":"新名字","Degree":"新学历"}'
          schema:
            type: string
      summary: 修改用户信息
      tags:
      - user
  /user/others:
    get:
      consumes:
      - application/json
      description: 得到其他用户的个人信息
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户邮箱
        in: formData
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":{"username":"kong"}}'
          schema:
            $ref: '#/definitions/user.InfoResponse'
      summary: 查看其他用户信息
      tags:
      - user
  /user/payment:
    post:
      consumes:
      - application/json
      description: 修改用户收款码
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: '{"message":"上传成功","data":map[string]interface{"url","path","sha"}}'
          schema:
            type: string
        "400":
          description: 上传失败,请检查token与其他配置参数是否正确
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: 上传收款码
      tags:
      - user
schemes:
- http
swagger: "2.0"